[S12P31A401-93] feat: Implement project list API

See merge request s12-final/S12P31A401!51

+++ spicoBackend/src/main/kotlin/com/ssafy/spico/domain/project/controller/ProjectController.kt
@@ -17,7 +17,7 @@ class ProjectController(
 ) {
     @GetMapping
     fun getProjects(
-        @RequestParam cursor: Long?,
+        @RequestParam cursor: Int?,
         @RequestParam size: Int,
         @RequestParam type: ProjectViewType,
     ): ApiResponse<List<ProjectResponse>> {
